PRO-2697: Renamed package to functional to avoid clashes.
