+++ src/functionalTest/java/uk/gov/hmcts/probate/functional/IntegrationTestBase.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.probate;
+package uk.gov.hmcts.probate.functional;
 
 import io.restassured.RestAssured;
 import net.thucydides.junit.spring.SpringIntegration;
@@ -8,7 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
-import uk.gov.hmcts.probate.util.TestUtils;
+import uk.gov.hmcts.probate.functional.util.TestUtils;
 
 @RunWith(SpringRunner.class)
 @ContextConfiguration(classes = TestContextConfiguration.class)
