+++ src/functionalTest/java/uk/gov/hmcts/probate/functional/TestContextConfiguration.java
@@ -1,11 +1,11 @@
-package uk.gov.hmcts.probate;
+package uk.gov.hmcts.probate.functional;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 
 @Configuration
-@ComponentScan("uk.gov.hmcts.probate")
+@ComponentScan("uk.gov.hmcts.probate.functional")
 @PropertySource("classpath:application.properties")
 public class TestContextConfiguration {
 }
