+++ src/functionalTest/java/uk/gov/hmcts/probate/functional/model/ClientAuthorizationResponse.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.probate.model;
+package uk.gov.hmcts.probate.functional.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
