+++ src/functionalTest/java/uk/gov/hmcts/probate/functional/serviceauth/SolCcdServiceServiceAuthTests.java
@@ -1,9 +1,9 @@
-package uk.gov.hmcts.probate.serviceauth;
+package uk.gov.hmcts.probate.functional.serviceauth;
 
 import net.serenitybdd.junit.runners.SerenityRunner;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import uk.gov.hmcts.probate.IntegrationTestBase;
+import uk.gov.hmcts.probate.functional.IntegrationTestBase;
 
 import static io.restassured.RestAssured.given;
